Updated the rom server to send any amount of roms to be saved in the rp2040 flash (memory is not boundary checked yet).
The emulator will be put into receiver mode if the button on GPIO 20 is held down on power up.
Once in receiver mode, the rom server can send roms over UART to be saved to flash.

Currently, the first program on the index is run on initialization, but programs will be selectable in the future.